+++ packages/health-check-nest/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.6.4
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.6.3
+
 ## 0.1.4
 
 ### Patch Changes